refactor: header 분리
